rename tooltipController -> selectionMenuController and fix imports
