+++ src/contentScript/textSelectionMenu/ActionRegistry.test.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import ActionRegistry from "@/contentScript/selectionTooltip/ActionRegistry";
+import ActionRegistry from "@/contentScript/textSelectionMenu/ActionRegistry";
 import { uuidv4 } from "@/types/helpers";
 
 describe("actionRegistry", () => {
