+++ src/tsconfig.strictNullChecks.json
@@ -321,13 +321,13 @@
     "./contentScript/pageEditor/beautify.ts",
     "./contentScript/popoverDom.ts",
     "./contentScript/ready.ts",
-    "./contentScript/selectionTooltip/ActionRegistry.test.ts",
-    "./contentScript/selectionTooltip/ActionRegistry.ts",
-    "./contentScript/selectionTooltip/SelectionToolbar.stories.tsx",
-    "./contentScript/selectionTooltip/SelectionToolbar.tsx",
-    "./contentScript/selectionTooltip/tooltipController.test.ts",
-    "./contentScript/selectionTooltip/tooltipController.tsx",
-    "./contentScript/selectionTooltip/useActionRegistry.ts",
+    "./contentScript/textSelectionMenu/ActionRegistry.test.ts",
+    "./contentScript/textSelectionMenu/ActionRegistry.ts",
+    "./contentScript/textSelectionMenu/SelectionToolbar.stories.tsx",
+    "./contentScript/textSelectionMenu/SelectionToolbar.tsx",
+    "./contentScript/textSelectionMenu/selectionMenuController.test.ts",
+    "./contentScript/textSelectionMenu/selectionMenuController.tsx",
+    "./contentScript/textSelectionMenu/useActionRegistry.ts",
     "./contentScript/setExtensionIdInApp.ts",
     "./contentScript/sidebarController.tsx",
     "./contentScript/sidebarDomControllerLite.ts",
